trim node cmd
