+++ cmd/node/Cargo.toml
@@ -13,32 +13,9 @@ ethrex-rlp = { workspace = true }
 ctor = "0.4.2"
 
 [dependencies]
-# ethrex
-ethrex-blockchain = { workspace = true }
-ethrex-common = { workspace = true }
-ethrex-l2-common = { workspace = true }
-ethrex-p2p = { workspace = true }
-ethrex-rpc = { workspace = true }
-ethrex-storage = { workspace = true }
-ethrex-storage-rollup = { workspace = true }
-ethrex-vm = { workspace = true }
-
-# mojave
 mojave-node-lib = { workspace = true }
 mojave-utils = { workspace = true }
-mojave-client = { workspace = true }
-mojave-signature = { workspace = true }
 
-axum = { workspace = true }
 clap = { workspace = true }
-hex = { workspace = true }
-futures = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
-tokio-stream = { workspace = true }
-tokio-util = { workspace = true }
-tower-http = { workspace = true, features = ["cors"] }
+tokio = { workspace = true }
 tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
