+++ cmd/node/src/cli.rs
@@ -1,5 +1,6 @@
-use clap::{Parser, Subcommand};
 use mojave_utils::options::Options;
+
+use clap::{Parser, Subcommand};
 use tracing::Level;
 
 #[allow(clippy::upper_case_acronyms)]
