--- cmd/sequencer/src/block_producer/mod.rs
@@ -1,7 +1,0 @@
-mod context;
-mod error;
-mod service;
-
-pub use context::BlockProducerContext;
-pub use error::BlockProducerError;
-pub use service::BlockProducer;
