+++ cmd/sequencer/src/main.rs
@@ -1,4 +1,3 @@
-pub mod block_producer;
 pub mod cli;
 pub mod error;
 pub mod rpc;
