+++ crates/node/src/lib.rs
@@ -1,3 +1,6 @@
 pub mod error;
 pub mod initializers;
+mod node;
 pub mod rpc;
+
+pub use node::Node;
